requirements.txt
Flask>=1.0.0
google-cloud-datastore==1.7.1
google-cloud-storage==1.13.0
google-cloud-logging==1.8.0
google-cloud-error_reporting==0.30.0
google-cloud-speech==1.0.0
google-auth==1.6.3
gunicorn==19.9.0
oauth2client==4.1.3
PyMySQL==0.9.2
six==1.11.0
requests[security]==2.20.1
psq==0.7.0
